Add capability to shift pitch with MIDI input
Create detectPitch function in PitchShifter class, which finds the
fundemental frequency of the input frame. Use the ratio between the
pitch of the input and the desired pitch of the output to determine
the pitch shift ratio. The desired pitch is set by the note value of
the midi input.

Add debug panel, with a text button listener that can display a set
value.

Put JuceHeader.h include in LiveSampler.h, so that only LiveSampler.h
needs to be included in every file.